refactor UserMenu component for readability
